Handle permanent redirection before feed fetching
This makes it possible to have permanent redirection handling and feed
fetching caching both work at a cost of an extra HEAD request.

Previously, if the response was "304 Not Modified", there was no way to
know whether there would be a permanent redirection on a request with no
If-Modified-Since/If-None-Match headers.