Mark imported feed as processed
This introduces an interesting edge case: if the feed was added and
feed2email was not run before running the migration script, the feed
will be considered as old and any entries fetched will be considered as
new and thus will be sent by email. But it is rather rare, it can happen
only once and it is not really harmful.