Handle feed redirection only when necessary
Before there was always a check before fetching the feed. After
monkey-patching open-uri to allow "http" to "https" scheme redirects, it
is possible to handle the redirection (to check if it is permanent) only
when there really is one. This should speed things up since now there is
only one HTTP request for feeds that do not redirect.