Pass correct (absolute) URI when building Entry
A potentially broken (relative) URI was passed and stored, but the check
of whether the entry was old was done with the fixed (absolute) URI. As
a result, if the two (original and fixed URIs) differed (e.g. relative
and absolute respectively), SQLite would throw an exception after
attempting to insert an entry with a URI that already existed.

TIL: SQL constraints are a very good idea.